skip location update if it's the same

+++ src/Models/SmartTimeTable.cs
@@ -5,7 +5,7 @@ namespace Models
     public class SmartTimeTable
     {
         //todo use dictionary and linked list for fast updates
-        public void Update(string routeId, int msBeforeArrival)
+        public void Update(string routeId, float msBeforeArrival)
         {
 
         }
